+++ src/app/mypage/recipe/page.tsx
@@ -3,5 +3,5 @@
 import { redirect } from 'next/navigation';
 
 export default function Recipe() {
-  redirect('/mypage/recipe/myRecipe');
+  redirect('/mypage/recipe/myRecipes');
 }
